Further separate global functionality
